+++ core/src/jsMain/kotlin/zakadabar/stack/frontend/builtin/titlebar/ZkAppHandle.kt
@@ -4,6 +4,8 @@
 package zakadabar.stack.frontend.builtin.titlebar
 
 import org.w3c.dom.events.Event
+import zakadabar.stack.frontend.application.ZkAppRouting
+import zakadabar.stack.frontend.application.application
 import zakadabar.stack.frontend.builtin.ZkElement
 import zakadabar.stack.frontend.builtin.button.ZkButton
 import zakadabar.stack.frontend.resources.ZkFlavour
@@ -14,9 +16,10 @@ import zakadabar.stack.frontend.util.plusAssign
  * The top-left part of the default layout, a button to close the menu and the title of the application.
  */
 open class ZkAppHandle(
-    private val appName: ZkElement,
-    private val onIconClick: (() -> Unit)? = null,
-    private val onTextClick: (() -> Unit)? = null
+    open val appName: ZkElement,
+    open val target: ZkAppRouting.ZkTarget? = null,
+    open val onIconClick: (() -> Unit)? = null,
+    open val onTextClick: (() -> Unit)? = null
 ) : ZkElement() {
 
     override fun onCreate() {
@@ -25,7 +28,13 @@ open class ZkAppHandle(
         + div(zkTitleBarStyles.appHandleButton) {
             + ZkButton(ZkIcons.notes, flavour = ZkFlavour.Custom, onClick = onIconClick)
         }
-        + appName.on("click") { onTextClick?.invoke() }
+        + appName.on("click") {
+            if (onTextClick != null) {
+                onTextClick?.invoke()
+            } else {
+                target?.let { application.changeNavState(it) }
+            }
+        }
 
         on("mousedown", ::onMouseDown)
     }