+++ site/src/jsMain/kotlin/zakadabar.site.frontend/pages/pages.kt
@@ -5,63 +5,60 @@ package zakadabar.site.frontend.pages
 
 import zakadabar.lib.markdown.frontend.MarkdownPage
 import zakadabar.lib.markdown.frontend.MarkdownPathPage
-import zakadabar.site.data.ContentQuery
 import zakadabar.stack.frontend.application.application
 import zakadabar.stack.frontend.builtin.titlebar.ZkAppTitle
 
+const val contentNamespace = "content"
+
 object Welcome : MarkdownPage(
-    "/${ContentQuery.dtoNamespace}/welcome/Welcome.md",
+    "/api/$contentNamespace/welcome/Welcome.md",
     SiteMarkdownContext("/Welcome", "welcome/")
 )
 
 object WhatsNew : MarkdownPage(
-    "/${ContentQuery.dtoNamespace}/WhatsNew.md",
+    "/api/$contentNamespace/WhatsNew.md",
     SiteMarkdownContext("/WhatsNew", "/")
 )
 
 object ShowCase : MarkdownPage(
-    "/${ContentQuery.dtoNamespace}/welcome/ShowCase.md",
+    "/api/$contentNamespace/welcome/ShowCase.md",
     SiteMarkdownContext("/ShowCase", "welcome/")
 )
 
 object Roadmap : MarkdownPage(
-    "/${ContentQuery.dtoNamespace}/Roadmap.md",
+    "/api/$contentNamespace/Roadmap.md",
     SiteMarkdownContext("/Roadmap", "/")
 )
 
 object GetStarted : MarkdownPage(
-    "/${ContentQuery.dtoNamespace}/GetStarted.md",
+    "/api/$contentNamespace/GetStarted.md",
     SiteMarkdownContext("/GetStarted", "/")
 )
 
 object GetHelp : MarkdownPage(
-    "/${ContentQuery.dtoNamespace}/help/GetHelp.md",
+    "/api/$contentNamespace/help/GetHelp.md",
     SiteMarkdownContext("/GetHelp", "help/")
 )
 
 
 object DocumentationIntro : MarkdownPage(
-    "/${ContentQuery.dtoNamespace}/help/Documentation.md",
+    "/api/$contentNamespace/help/Documentation.md",
     SiteMarkdownContext("/DocumentationIntro", "help/")
 )
 
 object FAQ : MarkdownPage(
-    "/${ContentQuery.dtoNamespace}/help/FAQ.md",
+    "/api/$contentNamespace/help/FAQ.md",
     SiteMarkdownContext("FAQ", "help/")
 )
 
 object Documentation : MarkdownPathPage() {
-
-    fun open(path: String) {
-        application.changeNavState(this, path)
-    }
 
     override fun setAppTitle() {
         application.title = ZkAppTitle("")
     }
 
     override fun url(): String {
-        return "/content/$path"
+        return "/api/$contentNamespace/$path"
     }
 
     override fun context() = SiteMarkdownContext(viewName, path)